Character property representations and lookup table
My main motivation in building this library is that I've been copying this
trie structure between a few different projects and so now wanted to put
it in its own library where I can reuse it and where others would hopefully
find it useful too.

Unfortunately I seem to have lost the code I wrote to generate this from
files in the Unicode character database, and so this is a snapshot of the
derived properties from Unicode 16.0.0 and unlikely to be updated again
for newer Unicode versions unless I one day summon the enthusiasm to
reimplement the trie generator.