Bring in the low-level state representation and various tests
As with earlier commits, this is mainly just adopting some code I've been
copying between other projects for a while and adapting for use as a
standalone library. This is adapted for the new organization of the
"properties" module but otherwise almost exactly the code I took from one
of those other codebases.